+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - Implementación de `LoadingIndicator` en `HomeScreen` para mostrar el estado de carga
 - Implementación de `LoadingIndicator` en `ProcedureScreen` para mostrar el estado de carga
 - Implementación de `LoadingIndicator` en `SettingScreen` para mostrar el estado de carga
+- Selección automática del primer espacio disponible al iniciar la aplicación
+
+### Removed
+- Errores de SonaarQube
 
 ### Fixed
 - Corrección de condiciones de carrera en login/logout